+++ packages/components/src/Chips/InternalChip.css.d.ts
@@ -1,0 +1,15 @@
+declare const styles: {
+  readonly "wrapper": string;
+  readonly "chip": string;
+  readonly "hasPrefix": string;
+  readonly "hasSuffix": string;
+  readonly "clickable": string;
+  readonly "input": string;
+  readonly "active": string;
+  readonly "invalid": string;
+  readonly "disabled": string;
+  readonly "button": string;
+  readonly "icon": string;
+};
+export = styles;
+