+++ packages/components/src/Menu/Menu.test.tsx
@@ -1,5 +1,11 @@
 import React from "react";
-import { cleanup, fireEvent, render, waitFor } from "@testing-library/react";
+import {
+  cleanup,
+  fireEvent,
+  render,
+  screen,
+  waitFor,
+} from "@testing-library/react";
 import { Menu } from ".";
 import { Button } from "../Button";
 
@@ -146,4 +152,31 @@ describe("Menu", () => {
     fireEvent.click(getByRole("button"));
     expect(clickHandler).toHaveBeenCalledTimes(1);
   });
+});
+
+it("should focus first action item from the menu when activated", () => {
+  const { getByRole } = render(
+    <Menu
+      activator={<Button label="Menu" />}
+      items={[
+        {
+          header: "Send as...",
+          actions: [
+            {
+              label: "Text Message",
+              icon: "sms",
+            },
+            {
+              label: "Email",
+              icon: "email",
+            },
+          ],
+        },
+      ]}
+    />,
+  );
+
+  fireEvent.click(getByRole("button"));
+  const firstMenuItem = screen.getAllByRole("menuitem")[0];
+  expect(firstMenuItem).toHaveFocus();
 });